backend/requirements.txt
aiosqlite==0.22.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.29.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.4
cryptography==42.0.8
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.8
google-auth==2.48.0
google-genai==0.6.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pillow==11.3.0
playwright==1.58.0
pluggy==1.6.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-playwright==0.7.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.31
starlette==0.45.3
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==14.2
wrapt==2.1.1
